Move mbql date filter utils out of DatePicker
